+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 
 ### Fixes
 
+- Fixed hardware wallets confirmation crashing issue when device is in "Busy" state ([PR 2522](https://github.com/input-output-hk/daedalus/pull/2522))
 - Fixed hardware wallets spending password issue ([PR 2519](https://github.com/input-output-hk/daedalus/pull/2519))
 
 ## 4.0.5